requirements.txt
# Web framework
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1

# ML and image processing
tensorflow==2.10.0
numpy==1.23.5
Pillow==9.3.0
opencv-python-headless==4.6.0.66

# Database
SQLAlchemy==1.4.27
python-dotenv==0.19.2

# Server
gunicorn==20.1.0

# Utilities
python-dateutil==2.8.2
requests==2.28.1 